requirements.txt
Jinja2==3.1.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
Pygments==2.15.1
PyYAML==6.0
rich==13.3.5
typing_extensions==4.5.0
